requirements.txt
flask
networkx
picamera2
numpy
opencv
paho-mqtt